+++ lib/msf/core/exploit/remote/http/kubernetes/client.rb
@@ -1,29 +1,13 @@
 # -*- coding: binary -*-
 
 require 'rex/proto/http/web_socket'
+require 'uri'
 
 module Msf
   class Exploit
     class Remote
       module HTTP
         module Kubernetes
-          module Error
-            class ApiError < ::StandardError
-            end
-
-            class AuthenticationError < ApiError
-            end
-
-            class UnexpectedStatusCode < ApiError
-              attr_reader :status_code
-
-              def initialize(status_code)
-                super
-                @status_code = status_code
-              end
-            end
-          end
-
           class Client
             USER_AGENT = 'kubectl/v1.22.2 (linux/amd64) kubernetes/8b5a191'.freeze
 
@@ -67,8 +51,8 @@ module Msf
             def exec_pod(name, namespace, command, options = {})
               options = {
                 'stdin' => false,
-                'stdout' => true,
-                'stderr' => true,
+                'stdout' => false,
+                'stderr' => false,
                 'tty' => false
               }.merge(options)
 
@@ -79,7 +63,7 @@ module Msf
                   {
                     'method' => 'GET',
                     'uri' => http_client.normalize_uri("/api/v1/namespaces/#{namespace}/pods/#{name}/exec"),
-                    'query' => build_query_string({
+                    'query' => URI.encode_www_form({
                       'command' => command,
                       'stdin' => !!options.delete('stdin'),
                       'stdout' => !!options.delete('stdout'),
@@ -198,22 +182,6 @@ module Msf
             EXEC_CHANNEL_RESIZE = 4
 
             attr_reader :http_client
-
-            def build_query_string(query_vars)
-              qary = []
-              query_vars.each_pair do |key, val|
-                key = key.to_s
-
-                key = Rex::Text.uri_encode(key)
-                if val.is_a? Array
-                  qary += val.map { |subval| "#{key}=#{Rex::Text.uri_encode(subval.to_s)}" }
-                else
-                  qary << "#{key}=#{Rex::Text.uri_encode(val.to_s)}"
-                end
-              end
-
-              qary.join('&')
-            end
 
             # TODO: Support receiving data directly as a table?
             # Accept: application/json;as=Table;g=meta.k8s.io;v=v1beta1
