+++ lib/msf/core/exploit/remote/http/kubernetes/error.rb
@@ -1,0 +1,28 @@
+# -*- coding: binary -*-
+
+module Msf
+  class Exploit
+    class Remote
+      module HTTP
+        module Kubernetes
+          module Error
+            class ApiError < ::StandardError
+            end
+
+            class AuthenticationError < ApiError
+            end
+
+            class UnexpectedStatusCode < ApiError
+              attr_reader :status_code
+
+              def initialize(status_code)
+                super
+                @status_code = status_code
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
